split out nat instance eip association
Terraform's `element` function was updated to throw exceptions when provided an empty list. Previously it just returned `nil` and the output was ignored for the nat instance/gateway scenarios.

This splits the eip association out into a separate resource that won't run for nat gateway situations removing the `element` evaluation issue.