route-table: Move route out of table resource
As mentioned in https://www.terraform.io/docs/providers/aws/r/route.html:

```
NOTE on Route Tables and Routes: Terraform currently provides both a standalone Route resource and a Route Table resource with routes defined in-line. At this time you cannot use a Route Table with in-line routes in conjunction with any Route resources. Doing so will cause a conflict of rule settings and will overwrite rules.
```

Even if the table id was outputted it would not be useable with the 
`route` resource. Each run would cause a “flapping” of the resource. One
run would add the route, the next would remove the route. 

Moving the route out of the table resource will prevent this conflict and
allow routes to be added in other modules.