subnets: Add custom tags per subnet type
This allows the application of user-defined tags to
the subnets created. The use case currently it to
tag subnets for Kubernetes clusters which need the
subnet tagged. Without setting this in TF the resources
would have to be ignored on each TF run or added
again manually after TF is run.